backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
sqlparse==0.5.3
